Re-organize API call related code into its own folder.
