+++ game_board/config.py
@@ -22,3 +22,13 @@ POINTS = {
     DIFFICULTY_LEVELS[1] : {'min':1, 'max':50},
     DIFFICULTY_LEVELS[2] : {'min':1, 'max':25}
 }
+
+
+# ========================================================================
+# Number of nodes given the difficulty
+# ========================================================================
+NODES = {
+    DIFFICULTY_LEVELS[0] : 5,
+    DIFFICULTY_LEVELS[1] : 10,
+    DIFFICULTY_LEVELS[2] : 20
+}
